+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Removed indexing of api-spec content (REST API specification)
+
 ## [3.4.2] - 2024-03-13
 
 ## [3.4.1] - 2024-02-13